doc: updates to MODELs and .NET README
Specifically, .NET's runner program made me realize that "count the
number of bytes in a match" is not really viable in some cases. For .NET
(and also, I imagine, Java), the most natural thing to do is count the
number of UTF-16 code units.

"counting bytes" is precisely equivalent to "count UTF-8 code units" in
cases where the haystack is valid UTF-8. There are some haystacks that
aren't valid UTF-8, so "counting bytes" really is the right thing there.
But conceptually speaking, it's fine to think of `count-spans` as
summing the lengths of matches in terms of UTF-8 or UTF-16 code units.