engines: add runner program for Perl
This also adds Perl to most of the curated benchmark definitions too. We
do hold off on adding it to any of the `compile` benchmarks because I
haven't been able to convinced myself that we aren't just measuring
Perl's regex compilation cache. See the README for more details on that.