cli: make 'rebar report' more permissive
It's not to uncommon (especially now, during development of rebar, where
things are in a state of flux) for old measurement CSV files to persist
even after benchmark definition names have changed. Previously, rebar
would just quit and refuse to do anything if it detected a measurement
with no corresponding benchmark definition. This commit changes that to
dropping the measurement and emitting a warning instead.

We also need to do an env_logger dance to set the default log level,
because we want these warnings shown by default without needing to set
RUST_LOG.