benchmarks: add many more 'test' benchmarks
This commit principally adds an entire host of "tests" masquerading as
benchmarks. Most of them focus on Unicode support, since that tends to
be where a lot of the subtle differences exist.

This isn't meant to be exhaustive, but it's a nice comparison point.
Practically speaking, it's also useful for confirming the behavior of a
regex engine is what you expect in the context of the rebar harness.
Namely, some regex engines make enabling Unicode support (*cough* Perl
*cough*) quite tricky. Still other regex engines don't support anything
outside of the BMP (*cough* .NET *cought*). These things can lead to
subtly different match counts in other benchmarks. So these "tests" can
help isolate where the differences might lay.