engines/java: remove from curated compilation benchmarks
The amazing speed at which Java is compiling regexes has lead me to be
very suspicious. Some regexes are being compiled in as little as 100ns
on my system, despite the fact that pattern compilation involves
recursive descent parsing and constructing a node tree.

This could very well be an incorrect decision, and I'm happy to revisit
upon advice from experts.