benchmarks: add ECMAScript engines to more benchmarks
They weren't being included in some benchmarks because of the use of
inline flags. Inline flags are often convenient because they let you
localize the effects of the flag to sub-expressions. So instead, we
delete some usages of flags that didn't seem to have an effect on the
result, and we replace others with work-arounds. For example, in
ECMAScript, instead of writing `(?s:.)`, we write `[\s\S]`. Lol.

We could add a 'dot-matches-new-line' flag like we have
'case-insensitive' and 'unicode' flags, but I'm not sure it's worth it.