cli: improve 'rebar build'
This makes the ergonomics and failure modes of 'rebar build' (hopefully)
a bit better. We do that by tweaking the presentation a little, but
principally by adding a new '[[engine.dependency]]' table. This table
lets you write commands that, if they fail, will cause 'rebar build' to
*not* build that regex engine and output an appropriate error message.
The dependency can also be made to fail if a regex search on the output
turns up no matches.

The idea here is to squash bad failure modes from tooling. For example,
running a Python 2 interpreter (probably unlikely these days) on Python
3 code is not likely to print a "hey you probably want Python 3" error
message. But 'rebar build' will do that---not perfectly---by saying that
'^Python 3.' doesn't match the output of 'python --version'. So I think
it will give a better clue.

The same idea applies to transitive dependencies. For example, 'cargo'
is obviously neede to build any of the runner programs written in Rust.
But many of them also need a C or C++ compiler.

And on and on.