readme: switch back to union for summary results
This is probably where I'll stay for now, since I think it's useful to
be able to get a purely global picture of relative performance, instead
of limiting oneself to the intersection.

The intersection makes much more sense for pairwise comparisons. But in
this sort of aggregate comparison, I think the union makes more sense.