cli: remove knowledge of benchmark model
In a previous revision of rebar, rebar bundled a number of regex engine
runner programs within itself. In that context, it made sense to bake in
knowledge about specific models, because it was already there.

But now? There's no real point in restricting the space of allowable
models. Instead, it's really up to the runner programs to recognize the
models. If they don't, then they (should) fail and this will in turn
translate to a measurement error that rebar will capture. Which is
totally fine.

However, there is still one piece of rebar that cares about the specific
model: whether to capture a haystack length or not. Without a haystack
length, rebar will report absolute timings instead of throughputs.
Absolute timings make much more sense for the `compile` and
`regex-redux` models, but not for any of the others.

It's not clear how rebar should "know" whether to record the haystack
length or not. It seems like we would need a separate model
configuration for that, but that seems like a lot of additional
infrastructure.

Instead of going out and building that, we let rebar know a little bit
about models: if it's the `compile` or `regex-redux` models, then it
doesn't capture haystack length. Otherwise, it does.

We can always revisit this later if more use cases arise for supporting
alternative models, or if more per-model configuration is needed.