cli: fix single pattern joining
When joining multiple patterns into a single one via an alternation, we
should enclose each pattern in a non-capturing group. This isn't
completely foolproof, but in order to break out of it, one has to write
a pattern that is invalid on its own.

This also fixes the `per-line` option so that the `literal`, `prepend`
and `append` options are correctly applied to each pattern.