benchmarks: remove non-ASCII from length-15 dictionary
The non-ASCII bytes cause the Rust regex crate to fail to compile the
regex unless Unicode mode is enabled. But it's just not worth making
this benchmark Unicode-aware for two words on the off-chance that it
has a real impact on results. (I don't believe it does.) So just drop
them.