cli: add '--row' flag for flipping the rows & columns
Since several new engines were added in the last few weeks, it nows
occurs quite often that there are just too many engines in the results
to compare. Namely, since engines are the columns, it doesn't take too
many of them to cause unreadable output in a terminal. This new flag
permits flipping the rows and columns, which is useful when one is only
comparing a few benchmarks.

It won't work in all cases, but it works in enough cases to be useful.