cli: rejigger filtering flags
Instead of making users awkwardly type a '!' to negate a pattern, we
instead add specific flags for negating patterns. Since we use lexopt
and thus have lower level and convenient control over the relationship
between flags, this is feasible to do. (It's harder, but possible I
believe, to do with Clap.)

So for example, -f foo is a whitelist rule while -F foo is a blacklist
rule.