cli: add new 'rebar rank' command
This makes it possible to compute the summary tables in the report on
the command line. This makes experimenting with summary rankings much
easier, and for example, lets one see summaries for any arbitrary slice
of benchmarks and also lets one experiment with how each aggregate
statistic changes the rankings.