cli: add 'rebar version' command and record it in measurements
Our measurements should definitely include the version of the rebar that
produced them. This is important for reproduction and report generation
purposes. Namely, report generation requires connecting a measurement
with a benchmark definition. Measurements are captured in time, but
definitions might change. So in order to produce a report for a
particular set of measurements, one should build an older version of
rebar. (Or at least use an older version of the benchmark definitions.)