cli: improve links in report generation
Basically, we ensure they are relative links that actually work.

We also try to include a link to the commit from which the report was
generated. This should hopefully make browsing links from the report
more accurate. Namely, the relative links, when looking at a specific
revision, should also be linking to other things within that revision.
This means that if we change the name of a haystack or regex file, the
links will all still work, assuming you're looking at the report in the
revision it was generated from.