cli: remove revision link
Well, it turns out that including a revision link the report output is
harder than I thought. The first problem, which is obvious in
retrospect, is that the revision won't include the report itself since
it was generated necessarily after the revision of the current rebar
tool. The second problem is that I don't know how to use a relative URL
to link to a specific commit revision. So if we wanted to do that, we'd
I think need to use an absolute URL.

For now we just write the revision and stop trying to be clever. If
folks want to see the state of the repo when the report was committed,
they'll have to navigate the history themselves. This shouldn't be too
bad, because once a report is generated, it's unlikely to be modified
later. So its commit history should be very short. It will therefore be
easy to go and find the revision in which it was introduced.

This commit also re-generates the report in the `record` directory so
that its links are correct now.