cli: tweak summary output
If the summary table(s) are empty, then don't show them at all.

Also, change the 'measure' report generation to exclude all engines from
the summary, and thus omit the summary section completely. The reason
for doing this is that the summary tables for ALL benchmarks are rather
misleading because there are many benchmarks with a small subset of
engines. Instead of just recklessly showing them, we don't do it at all.