scripts: record CPU information output
For now we just use 'lscpu' to do it, which is maybe a Linux thing. We
can add other mechanisms for other environments on an as-needed basis.

But the bottom line is that we should try to record environment
information that impacts interpretation of benchmark results. 'lscpu'
doesn't just record the CPU clock frequency for example, but it also
records what kinds of target features are available. Many optimized
regex engines use SIMD, so the target features can matter quite a bit.