engines: fix rust regex engines
While I'm working on regex-automata, I basically have a bunch of manual
`[patch.crates-io]` things to make crates use update code. It's a big
mess and a pain in the butt to deal with whenever `rust/regex` has a new
release.

Hopefully everything gets merged soon and we can get rid of all this
nonsense.