cli: re-work -t/--threshold flag
I always kind of hated how it worked. I had copied it from previous
related tools, such as 'cargo benchcmp' and 'critcmp'. But it's not
really what you want, because it doesn't obviously connect with the
speedup ratios actually shown in the output.

Instead, I made the thresholds filter directly on the speedup ratios
and turned it into a range.

I've also update the docs.

This works much better and is much more intuitive.