cli: make -d/--dir work more intuitively
Previously, the issue was that `cwd` in `benchmarks/engines.toml` was
being interpreted relative to the CWD of `rebar` itself. That in turn
means you can't really be anywhere and set --dir correctly, because some
part of the paths still depended on rebar's `cwd`.

This commit instead changes the `cwd` fields in `engines.toml` to be
specified relative to the benchmarks directory. That means that if you
have a path to the benchmarks directory (which is what --dir provides),
then you can derive everything else from that.