benchmarks: expunge rust/regex/meta
It's no longer used in any benchmark definitions in favor of rust/regex.

We do still keep the code and engine definition for it. It's useful to
have around if we need it. And it does actually expose other APIs that
wouldn't be fair to attribute to the top-level regex crate, such as
`search_with` which permits bypassing the synchronization overhead of a
`meta::Regex`'s internal `Cache` pool. The `regex` crate APIs don't have
anything equivalent to that.