regex: switch away from regex::bytes::Regex to regex::Regex
I admit it was more convenient to just run a Regex on the `&[u8]` from
the output of commands in some cases, but I don't think it's actually
necessary. We don't really expect, for example, the version output of a
command to contain invalid UTF-8. So we can just convert the entire
output to UTF-8 without much fuss.

The main motivation for doing this is that it should let us switch to
regex-lite in a subsequent commit, which should make compile times
faster.