deps: downgrade to toml 0.5.11
The dependency tree for toml 0.7 doesn't make me happy. It does looks
like you can disable some features, but parsing still requires
`toml_edit` which is where it looks like the bulk of the code is. It
also looks really hard to disable the `indexmap` dependency even though
it's technically optional? Dunno. It's pretty unclear to me.

Anyway, the toml 0.5 version works for me and shaves a couple of seconds
off of a clean build.