deps: drop direct dependency on atty
We still bring it in via env_logger. We could update env_logger, but
that would just replace it with the `is-terminal` crate, which brings in
a huge tree of dependencies. Hopefully env_logger will update to using
std::io::IsTerminal and then we'll be free of `atty` for good.