engines: fix versions
I knew I would forget to do this. When I bump versions, I need to update
the version of the crate itself to reflect it, because that's the
version that rebar uses. (There's probably a way to fix this by
spalunking through `cargo metadata`, but I couldn't quite figure it out
the last time I tried.)