engines/regress: update to 0.9.1
I had tried enabling `unicode_sets` when Unicode mode is enabled in a
benchmark definition, but this led to failures in the `03-date` curated
benchmark that I wasn't expecting.

It also seems that `\p{Greek}` is no longer valid, despite being valid
in `0.7.x`. Odd. So we remove `regress` from that benchmark.